requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
ipython==8.26.0
mysqlclient==2.2.4
PyJWT==2.9.0
python-decouple==3.8